Separate the diff() core function and the code to dump a diff.
The diff library is now renamed dump_diff. Its only client is the
pstore-diff tool and the two are separated purely to enable unit
testing of the dump_diff code. The diff() function moves into the core
library since it's now a key component of the export functionality.
