+++ include/pstore/exchange/export_names.hpp
@@ -46,8 +46,8 @@
 
 #include <unordered_map>
 
+#include "pstore/core/diff.hpp"
 #include "pstore/core/index_types.hpp"
-#include "pstore/diff/diff.hpp"
 #include "pstore/exchange/export_emit.hpp"
 
 namespace pstore {
@@ -94,7 +94,7 @@ namespace pstore {
                 auto names_index = index::get_index<trailer::indices::name> (db);
                 assert (generation > 0);
                 chunked_vector<address> container;
-                diff::diff (db, *names_index, generation - 1U, std::back_inserter (container));
+                diff (db, *names_index, generation - 1U, std::back_inserter (container));
                 emit_array (os, ind, std::begin (container), std::end (container),
                             [&names_index, &string_table, &db] (OStream & os1, indent ind1,
                                                                 address const addr) {
