+++ lib/CMakeLists.txt
@@ -48,7 +48,7 @@ add_subdirectory (broker)
 add_subdirectory (brokerface)
 add_subdirectory (command_line)
 add_subdirectory (core)
-add_subdirectory (diff)
+add_subdirectory (diff_dump)
 add_subdirectory (dump)
 add_subdirectory (exchange)
 add_subdirectory (http)
