+++ unittests/core/CMakeLists.txt
@@ -52,6 +52,7 @@ add_pstore_unit_test (pstore-core-unit-tests
     test_crc32.cpp
     test_database.cpp
     test_db_archive.cpp
+    test_diff.cpp
     test_generation_iterator.cpp
     test_hamt_map.cpp
     test_hamt_set.cpp
