+++ unittests/diff/CMakeLists.txt
@@ -44,12 +44,11 @@
 
 include (add_pstore)
 add_pstore_unit_test (pstore-diff-unit-tests
-    test_diff.cpp
     test_indices.cpp
     test_revision.cpp
 )
 target_link_libraries (pstore-diff-unit-tests
     PRIVATE
-        pstore-diff-lib
+        pstore-diff-dump
         pstore-unit-test-common
 )
